% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{filters_to_list}
\alias{filters_to_list}
\title{Convert GetEventFilters API return object to list object ready for POST call to GetDrawSheet API}
\usage{
filters_to_list(filters, up_list = list(), down_list = list())
}
\arguments{
\item{filters}{The filters object inside the response}

\item{up_list}{Any additional parameters to pass in (e.g. tournamentId)}

\item{down_list}{Used in the recursive calling of the function}
}
\value{
A list of lists, each containing filters required to call a Draw Sheet
}
\description{
Convert GetEventFilters API return object to list object ready for POST call to GetDrawSheet API
}
\examples{
filters_to_list(x$filters, list(
 tourType = x$tourType,
 tournamentId = x$tournamentId,
 weekNumber = 0
))
}
