% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{get_draw_sheet}
\alias{get_draw_sheet}
\title{Get a draw sheet using filters and GetDrawSheet API, and append darw sheet object to input parameters}
\usage{
get_draw_sheet(parameters_list)
}
\arguments{
\item{parameters_list}{List of filters to search for}
}
\value{
The same parameter list as passed in but with draw sheet object appended
}
\description{
Get a draw sheet using filters and GetDrawSheet API, and append darw sheet object to input parameters
}
\examples{
get_draw_sheet(list(
  tourType = "N",
  tournamentId = 1100044201,
  weekNumber = 0,
  drawsheetStructureCode: "KO",
  eventClassificationCode: "M",
  matchTypeCode: "S",
  playerTypeCode: "M"
))
}
