% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{get_event_filters}
\alias{get_event_filters}
\title{Get the event filters for a given tournament using GetEventFilters API}
\usage{
get_event_filters(key)
}
\arguments{
\item{key}{The tournament key as a string}
}
\value{
A nested list containing all allowed filters for a tournament
}
\description{
Get the event filters for a given tournament using GetEventFilters API
}
\examples{
get_event_filters("WC-ITFFS-AUS-02A-2019")
}
