% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{get_tournaments}
\alias{get_tournaments}
\title{Get a list of tournaments from TournamentApi}
\usage{
get_tournaments(circuit, start_date, days)
}
\arguments{
\item{circuit}{The circuit to query on: WCT, MT, WT, JT}

\item{start_date}{Date to take tournaments from (can be string)}

\item{days}{Number of days to include in query (integer)}
}
\value{
A dataframe listing or the return parameters from API
}
\description{
Get a list of tournaments from TournamentApi
}
\examples{
get_tournaments("WCT", "2020-01-01", 31)
}
