% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{response_type}
\alias{response_type}
\title{Test if server response is XML or JSON}
\usage{
response_type(resp)
}
\arguments{
\item{resp}{The server response as a string}
}
\value{
Lowercase string of either json or xml, otherwise empty string
}
\description{
Test if server response is XML or JSON
}
\examples{
response_type ("<xml></xml>")
}
